Update to initial configuration

+++ fabfile.py
@@ -5,8 +5,10 @@ import os
 import glob
 import json
 import datetime
+from StringIO import StringIO
 from fabric.api import env, local, run, sudo, runs_once, parallel, warn_only, cd, settings
 from fabric.operations import put, get
+from fabric.contrib.console import confirm
 
 # To debug communication issues un-comment the following
 # import logging
@@ -22,10 +24,7 @@ def _find_machines():
                      if not env.hosts or m["MachineName"] in env.hosts]
     env.hosts = [local("docker-machine ip {}".format(m["MachineName"]), capture=True) for m in machines
                  if not env.hosts or m["MachineName"] in env.hosts]
-    print(env.hosts)
     env.key_filename = [m["SSHKeyPath"] for m in machines]
-    # Use single key due to https://github.com/UCSC-Treehouse/pipelines/issues/5
-    # env.key_filename = "~/.ssh/id_rsa"
 
 
 _find_machines()
@@ -43,7 +42,8 @@ def cluster_up(count=1):
 		--azure-location "westus" \
 		--azure-ssh-user "ubuntu" \
 		--azure-open-port 80 \
-		--azure-size "Standard_D16_v3" \
+		--azure-size "Standard_D8_v3" \
+		--azure-storage-type "Standard_LRS" \
 		--azure-resource-group "bme230bstudents" \
               {}
               """.format(hostname))
@@ -53,21 +53,34 @@ def cluster_up(count=1):
 
 @parallel
 def configure_machines():
+    """ Push out class templates and download data sets """
     run("sudo usermod -aG docker ${USER}")
+
     put("class")
-    with cd("/mnt"):
-        run("wget -N https://bme230badmindiag811.blob.core.windows.net/data/tcga_target_gtex.h5") 
+
     run("sudo chown ubuntu:ubuntu /mnt")
+    run("mkdir -p /mnt/data /mnt/scratch")
+    with cd("/mnt/data"):
+        # run("wget -r -np -R 'index.html*' -N https://bme230badmindiag811.blob.core.windows.net/data/") 
+        run("wget -N https://bme230badmindiag811.blob.core.windows.net/data/tcga_target_gtex_train.h5") 
+        run("wget -N https://bme230badmindiag811.blob.core.windows.net/data/breast-cancer-wisconsin.data.csv") 
+        run("wget -N https://bme230badmindiag811.blob.core.windows.net/data/c2.cp.kegg.v6.1.symbols.gmt") 
+    run("sudo chown ubuntu:ubuntu /mnt")
+
+    run("""docker pull robcurrie/jupyter""")
     # run("chmod -R +r-w class/")
 
 
 @runs_once
 def cluster_down():
     """ Terminate ALL docker-machine machines """
-    for host in env.hostnames:
-        print("Terminating {}".format(host))
-        local("docker-machine stop {}".format(host))
-        local("docker-machine rm -f {}".format(host))
+    if confirm("Stop and delete all cluster machines?", default=False):
+        for host in env.hostnames:
+            print("Terminating {}".format(host))
+            with warn_only():
+                local("docker-machine stop {}".format(host))
+            with warn_only():
+                local("docker-machine rm -f {}".format(host))
 
 
 @runs_once
@@ -92,7 +105,8 @@ def jupyter_up():
 		-e NB_GID=`id -g` \
 		-p 80:8888 \
 		-v `echo ~`:/home/jovyan \
-		-v /mnt:/scratch \
+		-v /mnt/data:/home/jovyan/data \
+		-v /mnt/scratch:/scratch/home/jovyan/data \
 		robcurrie/jupyter start-notebook.sh \
 		--NotebookApp.password='sha1:c708a30ae9da:674e576d9a1b4c7fb421ea8b26e972cc63b93e59'
 	""", pty=False)
@@ -101,3 +115,16 @@ def jupyter_down():
     """ Shutdown the jupyter notebook server on each cluster machine """
     with warn_only():
         run("docker stop jupyter && docker rm jupyter")
+
+def backhaul():
+    """ Backhaul notebooks from all the cluster machines into /mnt/students """
+    username = env.host
+    with warn_only():
+        fd = StringIO()
+        paths = get("~/username.txt", fd)
+        if paths.succeeded:
+            username=fd.getvalue().strip()
+    print("username", username)
+    local("mkdir -p /mnt/students/{}".format(username))
+    with cd("/mnt/students/{}".format(username)):
+        get("/home/ubuntu/*", "/mnt/students/{}/".format(username))
